New gui.map module
